001-300/control-cam/backend/requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyserial==3.5
Werkzeug==3.1.3
